Fix the check if Integer or Long can be represented by float

+++ dataset/GitHub_Java/oracle.graal/truffle/src/com.oracle.truffle.api.interop/src/com/oracle/truffle/api/interop/DefaultIntegerExports.java
@@ -75,7 +75,7 @@ final class DefaultIntegerExports {
     static boolean fitsInFloat(Integer receiver) {
         int i = receiver;
         float f = i;
-        return (int) f == i;
+        return f == i;
     }
 
     @ExportMessage
